feat(editor): support download wallpaper as file
