+++ package.json
@@ -11,6 +11,7 @@
   "dependencies": {
     "core-js": "^3.6.5",
     "element-ui": "^2.15.6",
+    "file-saver": "^2.0.5",
     "vue": "^2.6.11",
     "vue-router": "^3.5.1"
   },
